requirements.txt
strands-agents
strands-agents-tools
boto3
python-dotenv
requests
scapy
Pillow
easyocr
